Fixes ui package
When you want to make a new package just right click on .thexgame folder and select new -> package
