+++ app/src/main/java/polymorphs/a301/f17/cs414/thexgame/ui/HistoryListAdapter.java
@@ -1,4 +1,4 @@
-package polymorphs.a301.f17.cs414.thexgame.polymorphs.a301.f17.cs414.thexgame.ui;
+package polymorphs.a301.f17.cs414.thexgame.ui;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -11,16 +11,15 @@ import java.util.List;
 import polymorphs.a301.f17.cs414.thexgame.R;
 
 /**
- * Created by thenotoriousrog on 10/13/17.
+ * Created by thenotoriousrog on 10/12/17.
  *
- * This class is another adapter to allow us to see games in the lists. Nothing special needs to happen in each of the items. They should just be strings.
+ * This is the adapter for the history adapter. It basically holds items and reports what is in it to the ListView
  */
 
-public class ActivityListAdapter extends ArrayAdapter<String> {
+public class HistoryListAdapter extends ArrayAdapter<String> {
 
-    public ActivityListAdapter(Context context, int resource, List<String> events)
-    {
-        super(context, resource, events);
+    public HistoryListAdapter(Context context, int resource, List<String> games){
+        super(context, resource, games);
     }
 
     // This populates each of the items in the ListView and controls things such as click actions and what not.
@@ -33,14 +32,13 @@ public class ActivityListAdapter extends ArrayAdapter<String> {
         {
             LayoutInflater vi;
             vi = LayoutInflater.from(getContext());
-            v = vi.inflate(R.layout.activity_item, null);
+            v = vi.inflate(R.layout.game_item, null);
         }
 
-        String eventItem = getItem(position); // grab the game string that we are working with.
+        String gameItem = getItem(position); // grab the game string that we are working with.
 
         // if we really wanted we could do things with this here. We could also send in list of objects not just string. Right we now we are not going to do anything.
 
         return v; // return the view to the ListView.
     }
-
 }
