+++ app/src/main/java/polymorphs/a301/f17/cs414/thexgame/ui/InGameUI.java
@@ -1,4 +1,4 @@
-package polymorphs.a301.f17.cs414.thexgame.polymorphs.a301.f17.cs414.thexgame.ui;
+package polymorphs.a301.f17.cs414.thexgame.ui;
 
 import android.app.Dialog;
 import android.content.Intent;
