+++ app/src/main/java/polymorphs/a301/f17/cs414/thexgame/ui/NotificationsUI.java
@@ -1,4 +1,4 @@
-package polymorphs.a301.f17.cs414.thexgame.polymorphs.a301.f17.cs414.thexgame.ui;
+package polymorphs.a301.f17.cs414.thexgame.ui;
 
 import android.app.Fragment;
 import android.os.Bundle;
@@ -10,8 +10,6 @@ import android.widget.ListView;
 import java.util.ArrayList;
 
 import polymorphs.a301.f17.cs414.thexgame.R;
-import polymorphs.a301.f17.cs414.thexgame.polymorphs.a301.f17.cs414.thexgame.ui.EventListAdapter;
-import polymorphs.a301.f17.cs414.thexgame.polymorphs.a301.f17.cs414.thexgame.ui.InvitationsListAdapter;
 
 /**
  * Created by thenotoriousrog on 10/13/17.
